Add "dynamic" switching of Kitty's color scheme.
The available color schemes are included in Kitty's config directory
inside the dotfile repository and have an extension *.Kcolor.
The included `setup` script allows choosing a color scheme as
```
./setup color <name>
```
If <name> is missing, all available schemes are displayed in the
terminal.
N.B. This is quite a wide display, requiring about 190 columns.